requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.76.0
anyio==4.12.1
async-timeout==5.0.1
asyncclick==8.3.0.7
attrs==25.4.0
av==16.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
firebase-messaging==0.4.5
frozenlist==1.8.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-crc32c==1.8.0
googleapis-common-protos==1.72.0
h11==0.16.0
http_ece==1.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
jiter==0.12.0
multidict==6.7.1
numpy>=1.24.0
oauthlib==3.3.1
openai==2.16.0
pillow==12.1.0
pydub>=0.25.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
ring-doorbell==0.9.13
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
